+++ app/page.tsx
@@ -1,7 +1,13 @@
+import Link from 'next/link';
+
 export default function Page() {
   return (
     <main>
-      <p>Hello</p>
+      <div>
+        <h1>Welcome to Bryggan</h1>
+        <Link href="/tidningen">Go to app</Link>
+        {/* {data.user == null ? <SignIn /> : <Link href="/tidningen">Hey {data.user.name}! Go to app</Link>} */}
+      </div>
     </main>
   );
 }
