+++ app/tidningen/error.tsx
@@ -1,0 +1,22 @@
+'use client';
+
+// Error components must be Client components
+import { useEffect } from 'react';
+
+type ErrorProps = { error: Error; reset: () => void };
+
+export default function Error({ error, reset }: ErrorProps) {
+  useEffect(() => {
+    // Log the error to an error reporting service
+    console.error(error);
+  }, [error]);
+
+  return (
+    <div>
+      <h1>Error</h1>
+      <p>{error.message}</p>
+      <p>The stack trace is:</p>
+      <pre>{error.stack}</pre>
+    </div>
+  );
+}
